# ENH: performance boost by vectorizing call to funeval