@oyamad 2017-11-22
(Added some commits and rebased.)

Many thanks, this is great!

* [Before](http://nbviewer.jupyter.org/github/QuantEcon/ContinuousDPs.jl/blob/a59c262c869e22d14ad2091c4fd6b81212f3e62c/examples/cdp_ex_optgrowth_jl.ipynb)
* [After](http://nbviewer.jupyter.org/github/QuantEcon/ContinuousDPs.jl/blob/f83f1df6db5d4f0359fbc691900dfec9779c8b6e/examples/cdp_ex_optgrowth_jl.ipynb)