@sglyon 2017-11-22
Looks good to me. Happy to see that performance boost!